Fix issue where local session state was not refreshed

+++ core/src/main/java/io/nextsense/android/base/db/objectbox/ObjectBoxDatabase.java
@@ -102,7 +102,7 @@ public class ObjectBoxDatabase implements Database {
     return deviceInternalStateBox.put(deviceInternalState);
   }
 
-  public LocalSession getLocalSession(int localSessionId) {
+  public LocalSession getLocalSession(long localSessionId) {
     return localSessionBox.get(localSessionId);
   }
